Renamed getNavigateToConfig to getSearchNavConfig and helpers to searchHelpers
fix
