+++ src/features/home/components/modules/OffersModule.tsx
@@ -7,7 +7,7 @@ import {
   OffersModule as OffersModuleType,
   RecommendedOffersModule,
 } from 'features/home/types'
-import { getNavigateToConfig } from 'features/navigation/SearchStackNavigator/helpers'
+import { getSearchNavConfig } from 'features/navigation/SearchStackNavigator/searchStackHelpers'
 import { OfferTileWrapper } from 'features/offer/components/OfferTile/OfferTileWrapper'
 import { useAdaptOffersPlaylistParameters } from 'libs/algolia/fetchAlgolia/fetchMultipleOffers/helpers/useAdaptOffersPlaylistParameters'
 import { analytics } from 'libs/analytics/provider'
@@ -70,7 +70,7 @@ export const OffersModule = (props: OffersModuleProps) => {
     locationParams,
     hitsPerPage: 20,
   }
-  const searchTabConfig = getNavigateToConfig('SearchResults', searchParams)
+  const searchTabConfig = getSearchNavConfig('SearchResults', searchParams)
 
   const moduleName = displayParameters.title ?? parameters?.title
