+++ src/features/internal/components/DeeplinksGeneratorForm.tsx
@@ -20,8 +20,8 @@ import {
   ScreensUsedByMarketing,
 } from 'features/internal/config/deeplinksExportConfig'
 import { getScreenPath } from 'features/navigation/RootNavigator/linking/getScreenPath'
-import { getSearchStackConfig } from 'features/navigation/SearchStackNavigator/helpers'
 import { isSearchStackScreen } from 'features/navigation/SearchStackNavigator/routes'
+import { getSearchStackConfig } from 'features/navigation/SearchStackNavigator/searchStackHelpers'
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
 import { isTabScreen } from 'features/navigation/TabBar/routes'
 import { MAX_PRICE_IN_CENTS } from 'features/search/helpers/reducer.helpers'
