+++ src/features/navigation/RootNavigator/Header/AccessibleTabBar.web.test.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { defaultDisabilitiesProperties } from 'features/accessibility/context/AccessibilityFiltersWrapper'
 import { useTabBarItemBadges } from 'features/navigation/helpers/useTabBarItemBadges'
 import * as navigationRefAPI from 'features/navigation/navigationRef'
-import { getSearchStackConfig } from 'features/navigation/SearchStackNavigator/helpers'
+import { getSearchStackConfig } from 'features/navigation/SearchStackNavigator/searchStackHelpers'
 import {
   DEFAULT_TAB_ROUTES,
   useTabNavigationContext,
