+++ src/features/offer/components/OfferHeader/OfferHeader.tsx
@@ -3,7 +3,7 @@ import { Animated } from 'react-native'
 import styled, { useTheme } from 'styled-components/native'
 
 import { FavoriteResponse, OfferResponseV2 } from 'api/gen'
-import { getSearchStackConfig } from 'features/navigation/SearchStackNavigator/helpers'
+import { getSearchStackConfig } from 'features/navigation/SearchStackNavigator/searchStackHelpers'
 import { useGoBack } from 'features/navigation/useGoBack'
 import { getShareOffer } from 'features/share/helpers/getShareOffer'
 import { WebShareModal } from 'features/share/pages/WebShareModal'
