+++ src/features/search/components/Buttons/FilterButton/FilterButton.native.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { navigate } from '__mocks__/@react-navigation/native'
-import { getNavigateToConfig } from 'features/navigation/SearchStackNavigator/helpers'
+import { getSearchNavConfig } from 'features/navigation/SearchStackNavigator/searchStackHelpers'
 import { SearchView } from 'features/search/types'
 import { userEvent, render, screen } from 'tests/utils'
 
@@ -12,19 +12,19 @@ jest.useFakeTimers()
 
 describe('FilterButton', () => {
   it('should contains the number of active filters', () => {
-    render(<FilterButton activeFilters={2} navigateTo={getNavigateToConfig(SearchView.Filter)} />)
+    render(<FilterButton activeFilters={2} navigateTo={getSearchNavConfig(SearchView.Filter)} />)
 
     expect(screen.getByText('2')).toBeOnTheScreen()
   })
 
   it('should not display badge when there are no active filters', () => {
-    render(<FilterButton activeFilters={0} navigateTo={getNavigateToConfig(SearchView.Filter)} />)
+    render(<FilterButton activeFilters={0} navigateTo={getSearchNavConfig(SearchView.Filter)} />)
 
     expect(screen.queryByTestId('searchFilterBadge')).not.toBeOnTheScreen()
   })
 
   it('should navigate with undefined params when pressing filter button', async () => {
-    render(<FilterButton activeFilters={1} navigateTo={getNavigateToConfig(SearchView.Filter)} />)
+    render(<FilterButton activeFilters={1} navigateTo={getSearchNavConfig(SearchView.Filter)} />)
 
     const filterButton = screen.getByTestId('searchFilterBadge')
     await user.press(filterButton)
@@ -37,7 +37,7 @@ describe('FilterButton', () => {
 
   describe('Accessibility', () => {
     it('should have an accessible label with the number of active filters', () => {
-      render(<FilterButton activeFilters={2} navigateTo={getNavigateToConfig(SearchView.Filter)} />)
+      render(<FilterButton activeFilters={2} navigateTo={getSearchNavConfig(SearchView.Filter)} />)
 
       expect(
         screen.getByLabelText('Voir tous les filtres\u00a0: 2 filtres actifs')
@@ -45,13 +45,13 @@ describe('FilterButton', () => {
     })
 
     it('should have an accessible label with one active filter', () => {
-      render(<FilterButton activeFilters={1} navigateTo={getNavigateToConfig(SearchView.Filter)} />)
+      render(<FilterButton activeFilters={1} navigateTo={getSearchNavConfig(SearchView.Filter)} />)
 
       expect(screen.getByLabelText('Voir tous les filtres\u00a0: 1 filtre actif')).toBeOnTheScreen()
     })
 
     it('should have an accessible label without active filter', () => {
-      render(<FilterButton activeFilters={0} navigateTo={getNavigateToConfig(SearchView.Filter)} />)
+      render(<FilterButton activeFilters={0} navigateTo={getSearchNavConfig(SearchView.Filter)} />)
 
       expect(screen.getByLabelText('Voir tous les filtres')).toBeOnTheScreen()
     })
