+++ src/features/search/helpers/useNavigateToSearch/useNavigateToSearch.ts
@@ -3,7 +3,7 @@ import { StackActions, useNavigation } from '@react-navigation/native'
 import { defaultDisabilitiesProperties } from 'features/accessibility/context/AccessibilityFiltersWrapper'
 import { DisabilitiesProperties } from 'features/accessibility/types'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
-import { getSearchStackConfig } from 'features/navigation/SearchStackNavigator/helpers'
+import { getSearchStackConfig } from 'features/navigation/SearchStackNavigator/searchStackHelpers'
 import { SearchStackRouteName } from 'features/navigation/SearchStackNavigator/types'
 import { SearchState } from 'features/search/types'
