+++ src/ui/components/buttons/SubcategoryButton/SubcategoryButton.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Platform, useWindowDimensions } from 'react-native'
 import styled from 'styled-components/native'
 
-import { getNavigateToConfig } from 'features/navigation/SearchStackNavigator/helpers'
+import { getSearchNavConfig } from 'features/navigation/SearchStackNavigator/searchStackHelpers'
 import { NativeCategoryEnum, SearchState } from 'features/search/types'
 import { useHandleFocus } from 'libs/hooks/useHandleFocus'
 import { useHandleHover } from 'libs/hooks/useHandleHover'
@@ -48,7 +48,7 @@ export const SubcategoryButton = ({
       {...hoverProps}
       onMouseDown={(e: Event) => e.preventDefault()} // Prevent focus on click
       onBeforeNavigate={onBeforeNavigate}
-      navigateTo={getNavigateToConfig('SearchResults', searchParams)}
+      navigateTo={getSearchNavConfig('SearchResults', searchParams)}
       testID={`SubcategoryButton ${label}`}
       accessibilityLabel={label}
       windowWidth={windowWidth}
